Extract ShotProfile and simplify ShotInformation
