+++ app/models/shot_information.rb
@@ -1,77 +1,16 @@
 # frozen_string_literal: true
 
 class ShotInformation < ApplicationRecord
-  extend Memoist
-
-  JSON_PROFILE_KEYS = %w[title author notes beverage_type steps tank_temperature target_weight target_volume target_volume_count_start legacy_profile_type type lang hidden reference_file changes_since_last_espresso version].freeze
+  include ShotProfile
 
   belongs_to :shot
 
-  def fahrenheit?
-    extra["enable_fahrenheit"].to_i == 1
-  end
-
-  def calculate_duration
-    index = if data["espresso_flow"]
-      [data["espresso_flow"].size, timeframe.size].min
-    else
-      timeframe.size
-    end
-    timeframe[index - 1].to_f
-  end
-
-  memoize def extra
+  def extra
     super.presence || {}
   end
 
-  memoize def profile_fields
+  def profile_fields
     super.presence || {}
-  end
-
-  memoize def tcl_profile_fields
-    profile_fields.except("json")
-  end
-
-  memoize def json_profile_fields
-    profile_fields["json"]
-  end
-
-  def tcl_profile
-    return if tcl_profile_fields.blank?
-
-    content = tcl_profile_fields.to_a.sort_by(&:first).map do |k, v|
-      v = "Visualizer/#{v}" if k == "profile_title"
-      v = "#{v.gsub(/Downloaded from Visualizer/, "").strip}\n\nDownloaded from Visualizer" if k == "profile_notes"
-      v = "{}" if v.blank?
-      v = "{#{v}}" if /\w\s\w/.match?(v)
-
-      "#{k} #{v}"
-    end
-
-    file_from_content(["#{shot.profile_title} from Visualizer", ".tcl"], content.join("\n"))
-  end
-
-  def json_profile
-    return if json_profile_fields.blank?
-
-    json = {}
-    JSON_PROFILE_KEYS.each do |key|
-      v = profile_fields["json"][key]
-      v = "Visualizer/#{v}" if key == "title"
-      v = "#{v.gsub(/Downloaded from Visualizer/, "").strip}\n\nDownloaded from Visualizer" if key == "notes"
-      json[key] = v
-    end
-
-    JSON.pretty_generate(json)
-  end
-
-  private
-
-  def file_from_content(filename, content)
-    file = Tempfile.new(filename)
-    file.write(content)
-    file.close
-    file.path
   end
 end
